feat(project) Refactored into one single CFN template.
* Updated README deploy instructions
* Cleared out unused project files
* Added buildkite pipeline
