+++ .gitignore
@@ -1,13 +1,12 @@
 .envrc
 handler.zip
-main
-lambdas/sfn/check-job
-lambdas/sfn/complete-job
-lambdas/sfn/submit-job
+/agent
+/check-job
+/complete-job
+/submit-job
 deploy.out.yml
 buildkite.zip
 coverage.txt
-buildkite.zip
 codebuid.out.yml
 buildkite.zip
 /vendor
