--- infra/codebuild/template.yml
@@ -1,93 +1,0 @@
----
-AWSTemplateFormatVersion: '2010-09-09'
-Description: 'codebuild: Codebuild Project which is used to run Buildkite jobs, a @wolfeidau template'
-Parameters:
-    EnvironmentName:
-      Type: String
-    EnvironmentNumber:
-      Type: String
-
-Resources:
-
-  SourceBucket:
-    Type: AWS::S3::Bucket
-
-  BuildkiteProject:
-    Type: AWS::CodeBuild::Project
-    DependsOn: BuildkiteCodeBuildRole
-    Properties:
-      Artifacts:
-        Type: no_artifacts
-      Environment:
-        ComputeType: BUILD_GENERAL1_SMALL
-        # using my custom codebuild docker image with buildkite installed along with golang
-        Image: wolfeidau/codebuild-docker-buildkite:17.09.0 
-        Type: LINUX_CONTAINER
-        EnvironmentVariables: 
-        - Name: ENVIRONMENT_NAME
-          Type: "PLAINTEXT"
-          Value: !Sub "${EnvironmentName}"
-        - Name: ENVIRONMENT_NUMBER
-          Type: "PLAINTEXT"
-          Value: !Sub "${EnvironmentNumber}"
-      Name: !Sub "BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}"
-      ServiceRole: !Ref BuildkiteCodeBuildRole
-      Source:
-        Type: S3
-        Location: !Sub "arn:aws:s3:::${SourceBucket}/buildkite.zip"
-
-  BuildkiteCodeBuildRole:
-    Type: AWS::IAM::Role
-    Properties:
-      AssumeRolePolicyDocument:
-        Version: "2012-10-17"
-        Statement:
-          Effect: Allow
-          Principal:
-            Service: codebuild.amazonaws.com
-          Action: sts:AssumeRole
-      Policies:
-      - PolicyName: BuildkiteSSMAccess
-        PolicyDocument:
-          Statement:
-          - Effect: Allow
-            Action: 'ssm:DescribeParameters'
-            Resource:
-            - '*'
-          - Effect: Allow
-            Action: 
-            - 'ssm:GetParameters'
-            - 'ssm:GetParameter'
-            Resource:
-            - !Sub "arn:aws:ssm:${AWS::Region}:${AWS::AccountId}:parameter/${EnvironmentName}/${EnvironmentNumber}/*"
-          - Effect: Allow
-            Action:
-            - logs:CreateLogGroup
-            - logs:CreateLogStream
-            - logs:PutLogEvents
-            Resource:
-            - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}"
-            - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}:*"
-          - Effect: Allow
-            Action:
-            - s3:PutObject
-            - s3:GetObject
-            - s3:GetObjectVersion
-            Resource:
-            - !Sub "arn:aws:s3:::codepipeline-${AWS::Region}-*"
-          - Effect: Allow
-            Action:
-            - s3:GetObject
-            - s3:GetObjectVersion
-            Resource:
-            - !Sub "arn:aws:s3:::${SourceBucket}/buildkite.zip"
-
-Outputs:
-  SourceBucket:
-    Value: !Ref SourceBucket
-    Export:
-      Name: !Sub "${AWS::StackName}-SourceBucket"
-  BuildkiteProject:
-    Value: !Ref BuildkiteProject
-    Export:
-      Name: !Sub "${AWS::StackName}-BuildkiteProject"
