--- infra/stepfunctions/template.yml
@@ -1,127 +1,0 @@
----
-AWSTemplateFormatVersion: '2010-09-09'
-Description: 'stepfunctions: State machine which monitors Buildkite codebuild jobs, a @wolfeidau template'
-Parameters:
-    EnvironmentName:
-      Type: String
-    EnvironmentNumber:
-      Type: String
-    SfnLambdaStack:
-      Description: 'Stack name of stack which deployed the step function job lambdas.'
-      Type: String
-
-Resources:
-
-  StateMachineCodebuildJobMonitor:
-    Type: 'AWS::StepFunctions::StateMachine'
-    Properties:
-      StateMachineName: !Sub "CodebuildJobMonitor-${EnvironmentName}-${EnvironmentNumber}"
-      DefinitionString: 
-        Fn::Sub:
-          - |
-            {
-              "Comment": "A state machine that submits a codebuild Job and monitors the Job until it completes.",
-              "StartAt": "Submit Job",
-              "States": {
-                "Submit Job": {
-                  "Type": "Task",
-                  "Resource": "${SfnSubmitLambdaARN}",
-                  "Next": "Wait X Seconds",
-                  "Retry": [
-                    {
-                      "ErrorEquals": [
-                        "States.ALL"
-                      ],
-                      "IntervalSeconds": 1,
-                      "MaxAttempts": 3,
-                      "BackoffRate": 2
-                    }
-                  ]
-                },
-                "Wait X Seconds": {
-                  "Type": "Wait",
-                  "SecondsPath": "$.wait_time",
-                  "Next": "Get Job Status"
-                },
-                "Get Job Status": {
-                  "Type": "Task",
-                  "Resource": "${SfnCheckLambdaARN}",
-                  "Next": "Job Complete?",
-                  "Retry": [
-                    {
-                      "ErrorEquals": [
-                        "States.ALL"
-                      ],
-                      "IntervalSeconds": 1,
-                      "MaxAttempts": 3,
-                      "BackoffRate": 2
-                    }
-                  ]
-                },
-                "Job Complete?": {
-                  "Type": "Choice",
-                  "Choices": [
-                    {
-                      "Variable": "$.build_status",
-                      "StringEquals": "FAILED",
-                      "Next": "Get Final Job Status"
-                    },
-                    {
-                      "Variable": "$.build_status",
-                      "StringEquals": "SUCCEEDED",
-                      "Next": "Get Final Job Status"
-                    }
-                  ],
-                  "Default": "Wait X Seconds"
-                },
-                "Get Final Job Status": {
-                  "Type": "Task",
-                  "Resource": "${SfnCompleteLambdaARN}",
-                  "End": true,
-                  "Retry": [
-                    {
-                      "ErrorEquals": [
-                        "States.ALL"
-                      ],
-                      "IntervalSeconds": 1,
-                      "MaxAttempts": 3,
-                      "BackoffRate": 2
-                    }
-                  ]
-                }
-              }
-            }
-          - SfnSubmitLambdaARN: !ImportValue
-              Fn::Sub: ${SfnLambdaStack}-SubmitJobFunctionArn
-            SfnCheckLambdaARN: !ImportValue
-              Fn::Sub: ${SfnLambdaStack}-CheckJobFunctionArn
-            SfnCompleteLambdaARN: !ImportValue
-              Fn::Sub: ${SfnLambdaStack}-CompleteJobFunctionArn
-      RoleArn: !GetAtt 'RoleCodebuildJobMonitor.Arn'
-
-  RoleCodebuildJobMonitor:
-    Type: 'AWS::IAM::Role'
-    Properties:
-      AssumeRolePolicyDocument:
-        Version: '2012-10-17'
-        Statement:
-        - Effect: Allow
-          Principal:
-            Service: !Sub 'states.${AWS::Region}.amazonaws.com'
-          Action: 'sts:AssumeRole'
-      Policies:
-      - PolicyName: lambda
-        PolicyDocument:
-          Statement:
-          - Effect: Allow
-            Action: 'lambda:InvokeFunction'
-            Resource:
-            - 'Fn::ImportValue': !Sub '${SfnLambdaStack}-SubmitJobFunctionArn'
-            - 'Fn::ImportValue': !Sub '${SfnLambdaStack}-CheckJobFunctionArn'
-            - 'Fn::ImportValue': !Sub '${SfnLambdaStack}-CompleteJobFunctionArn'
-
-Outputs:
-  StateMachineCodebuildJobMonitorArn:
-    Value: !Ref StateMachineCodebuildJobMonitor
-    Export:
-      Name: !Sub "${AWS::StackName}-StateMachineCodebuildJobMonitorArn"
