--- lambdas/agent-worker/deploy.sam.yml
@@ -1,58 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Description: 'lambda: Serverless Agent for Buildkite, a @wolfeidau template'
-Transform: AWS::Serverless-2016-10-31
-Parameters:
-    EnvironmentName:
-      Type: String
-    EnvironmentNumber:
-      Type: String
-    SfnCodebuildJobMonitorStack:
-      Type: String
-
-Resources:
-  Function:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: main
-      Timeout: 30
-      # For options see https://docs.aws.amazon.com/lambda/latest/dg/API_TracingConfig.html
-      Tracing: Active
-      Runtime: go1.x
-      CodeUri: ./handler.zip
-      Policies:
-        - AWSLambdaBasicExecutionRole
-        - AWSXrayWriteOnlyAccess
-        - Version: '2012-10-17' # Policy Document
-          Statement:
-            - Effect: Allow
-              Action:
-              - states:DescribeStateMachine
-              - states:StartExecution
-              - states:ListExecutions
-              Resource:
-              - 'Fn::ImportValue': !Sub '${SfnCodebuildJobMonitorStack}-StateMachineCodebuildJobMonitorArn'
-            - Effect: Allow
-              Action: 
-              - 'ssm:GetParameters'
-              - 'ssm:GetParameter'
-              - 'ssm:PutParameter'
-              Resource:
-              - !Sub "arn:aws:ssm:${AWS::Region}:${AWS::AccountId}:parameter/${EnvironmentName}/${EnvironmentNumber}/*"
-      Environment:   
-        Variables:
-          ENVIRONMENT_NAME:
-            Ref: EnvironmentName
-          ENVIRONMENT_NUMBER:
-            Ref: EnvironmentNumber
-          SFN_CODEBUILD_JOB_MONITOR_ARN:
-            'Fn::ImportValue': !Sub '${SfnCodebuildJobMonitorStack}-StateMachineCodebuildJobMonitorArn'
-      Events:
-        Timer:
-          Type: Schedule
-          Properties:
-            Schedule: rate(1 minute)
-Outputs:
-  FunctionArn:
-    Value: !GetAtt Function.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-FunctionArn"
