--- lambdas/sfn/deploy.sam.yml
@@ -1,121 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Description: 'lambda: Function which checks the status Buildkite jobs, a @wolfeidau template'
-Transform: AWS::Serverless-2016-10-31
-Parameters:
-    EnvironmentName:
-      Type: String
-    EnvironmentNumber:
-      Type: String
-
-Resources:
-  SfnFunctionRole:
-    Type: AWS::IAM::Role
-    Properties:
-      AssumeRolePolicyDocument: 
-        Version: "2012-10-17"
-        Statement: 
-          - 
-            Effect: "Allow"
-            Principal: 
-              Service: 
-                - "lambda.amazonaws.com"
-            Action: 
-              - sts:AssumeRole
-      ManagedPolicyArns:
-        - arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole
-        - arn:aws:iam::aws:policy/AWSXrayWriteOnlyAccess
-      Policies: 
-        - 
-          PolicyName: "BuildkiteSfnFunctionAccess"
-          PolicyDocument: 
-            Version: "2012-10-17"
-            Statement: 
-              - Effect: Allow
-                Action: 
-                - 'ssm:GetParameters'
-                - 'ssm:GetParameter'
-                Resource:
-                - !Sub "arn:aws:ssm:${AWS::Region}:${AWS::AccountId}:parameter/${EnvironmentName}/${EnvironmentNumber}/*"
-              - Effect: Allow
-                Action: 
-                - 'codebuild:BatchGetBuilds'
-                - 'codebuild:StartBuild'
-                Resource:
-                - !Sub "arn:aws:codebuild:${AWS::Region}:${AWS::AccountId}:project/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}"
-              - Effect: Allow
-                Action:
-                - logs:GetLogEvents
-                Resource:
-                - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}"
-                - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}:*"
-              - Effect: Allow
-                Action:
-                - logs:GetLogEvents
-                Resource:
-                - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}"
-                - !Sub "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/codebuild/BuildkiteProject-${EnvironmentName}-${EnvironmentNumber}:*"
-
-  SubmitJobFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: submit-job
-      Timeout: 30
-      # For options see https://docs.aws.amazon.com/lambda/latest/dg/API_TracingConfig.html
-      Tracing: Active
-      Runtime: go1.x
-      CodeUri: ./handler.zip
-      Role: !Sub ${SfnFunctionRole.Arn}
-      Environment:
-        Variables:
-          ENVIRONMENT_NAME:
-            Ref: EnvironmentName
-          ENVIRONMENT_NUMBER:
-            Ref: EnvironmentNumber
-
-  CheckJobFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: check-job
-      Timeout: 30
-      # For options see https://docs.aws.amazon.com/lambda/latest/dg/API_TracingConfig.html
-      Tracing: Active
-      Runtime: go1.x
-      CodeUri: ./handler.zip
-      Role: !Sub ${SfnFunctionRole.Arn}
-      Environment:
-        Variables:
-          ENVIRONMENT_NAME:
-            Ref: EnvironmentName
-          ENVIRONMENT_NUMBER:
-            Ref: EnvironmentNumber
-
-  CompleteJobFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: complete-job
-      Timeout: 30
-      # For options see https://docs.aws.amazon.com/lambda/latest/dg/API_TracingConfig.html
-      Tracing: Active
-      Runtime: go1.x
-      CodeUri: ./handler.zip
-      Role: !Sub ${SfnFunctionRole.Arn}
-      Environment:
-        Variables:
-          ENVIRONMENT_NAME:
-            Ref: EnvironmentName
-          ENVIRONMENT_NUMBER:
-            Ref: EnvironmentNumber
-
-Outputs:
-  SubmitJobFunctionArn:
-    Value: !GetAtt SubmitJobFunction.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-SubmitJobFunctionArn"
-  CheckJobFunctionArn:
-    Value: !GetAtt CheckJobFunction.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-CheckJobFunctionArn"
-  CompleteJobFunctionArn:
-    Value: !GetAtt CompleteJobFunction.Arn
-    Export:
-      Name: !Sub "${AWS::StackName}-CompleteJobFunctionArn"
