+++ pkg/agent/handler.go
@@ -1,4 +1,4 @@
-package handler
+package agent
 
 import (
 	"context"
