+++ pkg/registration/registration.go
